+++ R/draft.R
@@ -1,6 +1,6 @@
 #' Create a new CSAS document based on a template
 #'
-#' Create a draft of an R Markdown CSAS document
+#' Create a draft of an R Markdown CSAS document.
 #'
 #' @param type The type of document to draft. Must be one of `resdoc`, `fsar`,
 #' `sr`, or `techreport`.
@@ -14,11 +14,25 @@
 #' function for further details.
 #' @importFrom cli cli_alert_success
 #'
+#' @seealso [render()]
+#' @returns A new project folder with template files for creating a csasdown
+#' report.
 #' @examples
-#' \dontrun{
+#' \donttest{
+#' # create a temporary example folder:
+#' wd <- getwd()
+#' example_path <- file.path(tempdir(), "csasdown-example")
+#' dir.create(example_path)
+#' setwd(example_path)
+#'
+#' # create a draft template:
 #' csasdown::draft("resdoc")
-#' csasdown::draft("fsar")
-#' csasdown::draft("techreport")
+#'
+#' # return to original working directory after running example:
+#' setwd(wd)
+#'
+#' # clean up:
+#' unlink(example_path, recursive = TRUE, force = TRUE)
 #' }
 #' @export
 draft <- function(